+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
+- K8s.Operation.put_label_selector/2
+- K8s.Operation.get_label_selector/1
 - Per connection http provider configuration
 
 ### Changed